adjust typesetting of sexagesimal formats
Modify the way that the sexagesimal formats are presented in the
text: writing "hhmmss" is hard to understand, though the examples
do clarify this.  I changed the fonts of the placeholder and literal
parts of these formats which at least means they are distinguishable
if you look closely.  If the authors/editor prefer to indicate this
in some other way, feel free to change it.

I also fixed the opening quotes.